requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.6.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
model-bakery==1.20.5
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytest-factoryboy==2.8.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
tzdata==2025.2
